node loader, ssr, & csr updates
