+++ christians-lawyer/next.config.mjs
@@ -2,6 +2,14 @@
 const nextConfig = {
     experimental: {
         reactCompiler: true
+    },
+    webpack: (config) => {
+        config.module.rules.push({
+            test: /\.node$/,
+            use: [{ loader: "nextjs-node-loader" }]
+        });
+
+        return config;
     }
 };
